feat(zakaz): korobka/pachka soni va hajmi o'nlik (kasr) son qabul qiladi
Miqdor (quantity) allaqachon Decimal edi; korobka/pachka soni faqat int (z.int,
Int ustun, mobil inputMode=numeric — nuqtasiz) edi. Og'irlik/hajm bilan sotiladigan
tovarlar uchun kasr kerak (1.5, 0.5).

- Schema: Product.packSize + PurchaseOrderItem.packSize/packCount → Decimal(20,3)
  (migratsiya 20260615120719_pack_decimal; Int→Decimal widening, ma'lumot saqlanadi;
  trgm DROP olib tashlandi/tiklandi).
- Validatsiya: pack* dan .int() olib tashlandi (sotib-olish + taminotchilar).
- O'qishlar Number() bilan o'raldi (supplierItems, profil, order-detail, PDF).
- Inputlar: blok/pachka inputMode=decimal; LeadTimeEditor pack int:false.

+++ src/app/(app)/baza/taminotchilar/actions.ts
@@ -303,7 +303,7 @@ export async function createSupplierAction(
 const skuParamSchema = z.object({
   productId: z.coerce.number().int().positive(),
   leadTimeDays: z.coerce.number().int().min(0).max(365).nullable().optional(),
-  packSize: z.coerce.number().int().positive().max(100_000).nullable().optional(),
+  packSize: z.coerce.number().positive().max(100_000).nullable().optional(),
   purchasePrice: z.coerce.number().nonnegative().max(1_000_000_000_000).nullable().optional(),
 });
